<refactor>[UserBalance]: <Added Balance Controller>
