+++ paydev/src/main/java/dev/paydev/paydev/controllers/UserController.java
@@ -1,31 +1,25 @@
 package dev.paydev.paydev.controllers;
 
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import dev.paydev.paydev.domain.ViewModels.UserBalanceViewModel;
-import dev.paydev.paydev.domain.user.User;
-import dev.paydev.paydev.repository.services.BalanceService;
-import dev.paydev.paydev.repository.services.UserService;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import dev.paydev.paydev.domain.user.User;
+import dev.paydev.paydev.repository.services.UserService;
 
 
 @RestController
 @RequestMapping("/api/v1/user")
 public class UserController {
     private final UserService _userService;
-    private final BalanceService _balanceService;
 
-    public UserController(UserService userService, BalanceService balanceService) {
+    public UserController(UserService userService) {
         this._userService = userService;
-        this._balanceService = balanceService;
     }
 
     @GetMapping("/{id}")
@@ -38,18 +32,6 @@ public class UserController {
     public ResponseEntity<User> createUser(@RequestBody User user) {
         User userCreated = _userService.createUser(user);
         return ResponseEntity.ok(userCreated);
-    }
-
-    @GetMapping("balance/{id}")
-    public ResponseEntity<Object> getUserBalance(@PathVariable Long id) {
-        Object userBalance = _balanceService.GetUserBalance(id);
-        return ResponseEntity.ok(userBalance);
-    }
-
-    @PutMapping("balance")
-    public ResponseEntity<User> addUserBalance(@RequestBody UserBalanceViewModel userBalance) {
-        User userUpdated = _balanceService.updateUserBalance(userBalance.getUserId(), userBalance.getAmmount());
-        return ResponseEntity.ok(userUpdated);
     }
 
     @DeleteMapping("/{id}")
